+++ spec/controllers/user_controller_spec.rb
@@ -56,6 +56,234 @@ describe UserController do
 
   end
 
+  describe 'GET confirm' do
+
+    context 'if the post redirect cannot be found' do
+
+      it 'renders bad_token' do
+        get :confirm, :email_token => ''
+        expect(response).to render_template(:bad_token)
+      end
+
+    end
+
+    context 'the post redirect circumstance is login_as' do
+
+      before :each do
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect =
+          PostRedirect.
+            create(:uri => '/', :user => @user, :circumstance => 'login_as')
+
+        get :confirm, :email_token => @post_redirect.email_token
+      end
+
+      it 'confirms the post redirect user' do
+        expect(@user.reload.email_confirmed).to eq(true)
+      end
+
+      it 'logs in as the post redirect user' do
+        expect(session[:user_id]).to eq(@user.id)
+      end
+
+      it 'sets the user_circumstance to login_as' do
+        expect(session[:user_circumstance]).to eq('login_as')
+      end
+
+      it 'redirects to the post redirect uri' do
+        expect(response).to redirect_to('/?post_redirect=1')
+      end
+
+    end
+
+    context 'the post redirect circumstance is change_password' do
+
+      before :each do
+        @user = FactoryGirl.create(:user)
+        @post_redirect =
+          PostRedirect.create(:uri => edit_password_change_path,
+                              :user => @user,
+                              :circumstance => 'change_password')
+
+        get :confirm, :email_token => @post_redirect.email_token
+      end
+
+      it 'sets the change_password_post_redirect_id session key' do
+         expect(session[:change_password_post_redirect_id]).
+           to eq(@post_redirect.id)
+      end
+
+      it 'does not log the user in' do
+        expect(session[:user_id]).to eq(nil)
+      end
+
+      it 'logs out a user who does not own the post redirect' do
+        logged_in_user = FactoryGirl.create(:user)
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect =
+          PostRedirect.create(:uri => edit_password_change_path,
+                              :user => @user,
+                              :circumstance => 'change_password')
+
+        session[:user_id] = logged_in_user.id
+        get :confirm, :email_token => @post_redirect.email_token
+
+        expect(session[:user_id]).to be_nil
+      end
+
+      it 'does not log out a user if they own the post redirect' do
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect =
+          PostRedirect.create(:uri => edit_password_change_path,
+                              :user => @user,
+                              :circumstance => 'change_password')
+
+        session[:user_id] = @user.id
+        get :confirm, :email_token => @post_redirect.email_token
+
+        expect(session[:user_id]).to eq(@user.id)
+        expect(assigns[:user]).to eq(@user)
+      end
+
+      it 'does not confirm an unconfirmed user' do
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect =
+          PostRedirect.create(:uri => edit_password_change_path,
+                              :user => @user,
+                              :circumstance => 'change_password')
+
+        get :confirm, :email_token => @post_redirect.email_token
+
+        expect(@user.reload.email_confirmed).to eq(false)
+      end
+
+      it 'sets the user_circumstance to change_password' do
+        expect(session[:user_circumstance]).to eq('change_password')
+      end
+
+      it 'redirects to the post redirect uri' do
+        expect(response).
+          to redirect_to('/profile/change_password?post_redirect=1')
+      end
+
+    end
+
+    context 'if the currently logged in user is an admin' do
+
+      before :each do
+        @admin = FactoryGirl.create(:user, :admin_level => 'super')
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
+
+        session[:user_id] = @admin.id
+        get :confirm, :email_token => @post_redirect.email_token
+      end
+
+      it 'does not confirm the post redirect user' do
+        expect(@user.reload.email_confirmed).to eq(false)
+      end
+
+      it 'stays logged in as the admin user' do
+        expect(session[:user_id]).to eq(@admin.id)
+      end
+
+      it 'sets the user_circumstance to normal' do
+        expect(session[:user_circumstance]).to eq('normal')
+      end
+
+      it 'redirects to the post redirect uri' do
+        expect(response).to redirect_to('/?post_redirect=1')
+      end
+
+    end
+
+    context 'if the currently logged in user is not an admin and owns the post redirect' do
+
+      before :each do
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
+
+        session[:user_id] = @user.id
+        get :confirm, :email_token => @post_redirect.email_token
+      end
+
+      it 'confirms the post redirect user' do
+        expect(@user.reload.email_confirmed).to eq(true)
+      end
+
+      it 'stays logged in as the user' do
+        expect(session[:user_id]).to eq(@user.id)
+      end
+
+      it 'sets the user_circumstance to normal' do
+        expect(session[:user_circumstance]).to eq('normal')
+      end
+
+      it 'redirects to the post redirect uri' do
+        expect(response).to redirect_to('/?post_redirect=1')
+      end
+
+    end
+
+    context 'if the currently logged in user is not an admin and does not own the post redirect' do
+
+      before :each do
+        @current_user = FactoryGirl.create(:user)
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
+
+        session[:user_id] = @current_user.id
+        get :confirm, :email_token => @post_redirect.email_token
+      end
+
+      it 'confirms the post redirect user' do
+        expect(@user.reload.email_confirmed).to eq(true)
+      end
+
+      # FIXME: There's no reason this should be allowed
+      it 'gets logged in as the post redirect user' do
+        expect(session[:user_id]).to eq(@user.id)
+      end
+
+      it 'sets the user_circumstance to normal' do
+        expect(session[:user_circumstance]).to eq('normal')
+      end
+
+      it 'redirects to the post redirect uri' do
+        expect(response).to redirect_to('/?post_redirect=1')
+      end
+
+    end
+
+    context 'if there is no logged in user' do
+
+      before :each do
+        @user = FactoryGirl.create(:user, :email_confirmed => false)
+        @post_redirect = PostRedirect.create(:uri => '/', :user => @user)
+
+        get :confirm, :email_token => @post_redirect.email_token
+      end
+
+      it 'confirms the post redirect user' do
+        expect(@user.reload.email_confirmed).to eq(true)
+      end
+
+      it 'gets logged in as the post redirect user' do
+        expect(session[:user_id]).to eq(@user.id)
+      end
+
+      it 'sets the user_circumstance to normal' do
+        expect(session[:user_circumstance]).to eq('normal')
+      end
+
+      it 'redirects to the post redirect uri' do
+        expect(response).to redirect_to('/?post_redirect=1')
+      end
+
+    end
+
+  end
+
 end
 
 # TODO: Use route_for or params_from to check /c/ links better
@@ -514,76 +742,6 @@ describe UserController, "when sending another user a message" do
     expect(mail.body).to include("Just a test!")
     #mail.to_addrs.first.to_s.should == users(:silly_name_user).name_and_email # TODO: fix some nastiness with quoting name_and_email
     expect(mail.from_addrs.first.to_s).to eq(users(:bob_smith_user).email)
-  end
-
-end
-
-describe UserController, "when changing password" do
-  render_views
-
-  it "should show the email form when not logged in" do
-    get :signchangepassword
-    expect(response).to render_template('signchangepassword_send_confirm')
-  end
-
-  it "should send a confirmation email when logged in normally" do
-    session[:user_id] = users(:bob_smith_user).id
-    get :signchangepassword
-    expect(response).to render_template('signchangepassword_confirm')
-
-    deliveries = ActionMailer::Base.deliveries
-    expect(deliveries.size).to  eq(1)
-    mail = deliveries[0]
-    expect(mail.body).to include("Please click on the link below to confirm your email address")
-  end
-
-  it "should send a confirmation email when have wrong login circumstance" do
-    session[:user_id] = users(:bob_smith_user).id
-    session[:user_circumstance] = "bogus"
-    get :signchangepassword
-    expect(response).to render_template('signchangepassword_confirm')
-  end
-
-  it "should show the password change screen when logged in as special password change mode" do
-    session[:user_id] = users(:bob_smith_user).id
-    session[:user_circumstance] = "change_password"
-    get :signchangepassword
-    expect(response).to render_template('signchangepassword')
-  end
-
-  it "should change the password, if you have right to do so" do
-    session[:user_id] = users(:bob_smith_user).id
-    session[:user_circumstance] = "change_password"
-
-    old_hash = users(:bob_smith_user).hashed_password
-    post :signchangepassword, { :user => { :password => 'ooo', :password_confirmation => 'ooo' },
-                                :submitted_signchangepassword_do => 1
-                                }
-    expect(users(:bob_smith_user).reload.hashed_password).not_to eq(old_hash)
-
-    expect(response).to redirect_to(:controller => 'user', :action => 'show', :url_name => users(:bob_smith_user).url_name)
-  end
-
-  it "should not change the password, if you're not logged in" do
-    session[:user_circumstance] = "change_password"
-
-    old_hash = users(:bob_smith_user).hashed_password
-    post :signchangepassword, { :user => { :password => 'ooo', :password_confirmation => 'ooo' },
-                                :submitted_signchange_password => 1
-                                }
-    expect(users(:bob_smith_user).hashed_password).to eq(old_hash)
-  end
-
-  it "should not change the password, if you're just logged in normally" do
-    session[:user_id] = users(:bob_smith_user).id
-    session[:user_circumstance] = nil
-
-    old_hash = users(:bob_smith_user).hashed_password
-    post :signchangepassword, { :user => { :password => 'ooo', :password_confirmation => 'ooo' },
-                                :submitted_signchange_password => 1
-                                }
-
-    expect(users(:bob_smith_user).hashed_password).to eq(old_hash)
   end
 
 end